Fix tmux options
Setting TERM to tmux-xterm caused issues with line breaks in centos 7.5
(bash 4.2) w/ tmux and kitty. I was setting the term due to previous
issues with nvim colours in tmux. Right now my setup is differnt:
lazyvim handles the color scheme and I use kitty's ssh binding for ssh,
which sets xterm to xterm-kitty and seems to fix the issue in the centos
servers.